backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
SQLAlchemy
pydantic
python-dotenv
pymysql
reportlab
openpyxl
apscheduler
